fixed the type of second parameter of memset()

+++ include/string.h
@@ -5,7 +5,7 @@
 
 char *itoa(int);
 void memcpy(void *, const void *, size_t);
-void memset(void *, int, size_t);
+void memset(void *, uint8_t, size_t);
 size_t strlen(const char *);
 void strcpy(char *d, const char *s);
